+++ pageserver/src/config.rs
@@ -4,6 +4,8 @@
 //! file, or on the command line.
 //! See also `settings.md` for better description on every parameter.
 
+pub mod ignored_fields;
+
 use std::env;
 use std::num::NonZeroUsize;
 use std::sync::Arc;
@@ -206,12 +208,6 @@ pub struct PageServerConf {
 
     /// When set, include visible layers in the next uploaded heatmaps of an unarchived timeline.
     pub generate_unarchival_heatmap: bool,
-}
-
-/// List of items in the on-disk config file that we ignored when deserializing into
-/// [`pageserver_api::config::ConfigToml`].
-pub struct IgnoredConfigItems {
-    pub paths: Vec<String>,
 }
 
 /// Token for authentication to safekeepers
